requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7
